requirements.txt
absl-py==2.1.0
aiobotocore
aiohttp
aioitertools
aiosignal
alabaster
altair
anaconda-anon-usage
anaconda-catalogs
anaconda-client
anaconda-cloud-auth
anaconda-navigator
anaconda-project
anyio
appdirs==1.4.4
archspec
argon2-cffi
argon2-cffi-bindings
arrow
asgiref==3.8.1
astroid
astropy
asttokens
astunparse==1.6.3
async-lru
atomicwrites==1.4.0
attrs
Automat
autopep8
Babel
backports.functools-lru-cache
backports.tempfile
backports.weakref==1.0.post1
bcrypt
beautifulsoup4
binaryornot
black
bleach
blinker
bokeh
boltons
botocore
Bottleneck
Brotli
cachelib==0.13.0
cachetools
certifi
cffi
chardet
charset-normalizer
click
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle
clyent==1.2.2
colorama
colorcet
comm
conda
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli==1.0.75
conda-token
conda-verify==3.4.2
conda_index
conda_package_streaming
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
dask
datashader
debugpy
decorator
defusedxml
diff-match-patch
diffusers @ git+https://github.com/huggingface/diffusers.git@9366c8f84bfe47099ff047272661786ebb54721d
dill
distributed
distro
Django==4.2.3
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.14.0
dnspython==2.7.0
docker==7.1.0
docstring-to-markdown
docutils
entrypoints
et-xmlfile==1.1.0
executing
facebook-sdk==3.1.0
fastapi==0.115.6
fastjsonschema
filelock
flake8
Flask
Flask-Session==0.8.0
flatbuffers==24.12.23
fonttools==4.25.0
frozenlist
fsspec
ftfy==6.3.0
future
gast==0.6.0
gensim
geopandas==1.0.1
gitdb
GitPython
gmpy2
google-ai-generativelanguage==0.6.15
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn
h11==0.14.0
h5py==3.12.1
HeapDict
holoviews
httplib2==0.22.0
httptools==0.6.4
huggingface_hub @ git+https://github.com/huggingface/huggingface_hub.git@aca9aec0e4493897ed74af8051a33e62a6fd1e2c
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
incremental
inflection==0.5.1
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jedi
jellyfish
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keras==3.7.0
keyring
kiwisolver
lazy-object-proxy
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libclang==18.1.1
libmambapy
linkify-it-py
llvmlite
lmdb
locket
lxml
lz4
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
menuinst
mistune
mkl-fft
mkl-random
mkl-service==2.4.0
ml-dtypes==0.4.1
more-itertools
mpmath
msgpack
msgspec==0.19.0
multidict
multipledispatch
munkres==1.1.4
mypy
mypy-extensions
namex==0.0.8
navigator-updater
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
opencv-python==4.5.5.64
openpyxl==3.0.10
opt_einsum==3.4.0
optree==0.13.1
Orange3-Associate==1.3.0
overrides
packaging
pandas==2.2.3
pandocfilters
panel
param
paramiko
parsel
parso
partd
pathlib
pathspec
patsy==0.5.3
pexpect
pickleshare
Pillow==9.5.0
pkce
pkginfo
platformdirs
plotly
pluggy
ply==3.11
prometheus-client
prompt-toolkit
Protego
proto-plus==1.25.0
protobuf==5.29.3
psutil
psycopg2==2.9.9
ptyprocess
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules==0.2.8
pycodestyle
pycosat
pycparser
pyct
pycurl==7.45.2
pydantic
pydeck
PyDispatcher==2.0.5
pydocstyle
pyerfa
pyflakes
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder==0.4.0
pymongo==4.11.1
PyNaCl
pyodbc
pyogrio==0.10.0
pyOpenSSL
pyparsing
pyproj==3.7.0
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytest
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc==1.0.0
python-lsp-server
python-multipart==0.0.20
python-slugify
python-snappy
pytoolconfig
pytz
pyviz_comms
pywavelets
pywin32==305.1
pywin32-ctypes
pywinpty
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
referencing
regex
requests
requests-file
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rope
rpds-py
rsa==4.9
Rtree
ruamel-yaml-conda
ruamel.yaml
s3fs
safetensors==0.4.5
scikit-image
scikit-learn
scipy==1.11.4
Scrapy
seaborn
semver
Send2Trash
service-identity
shapely==2.0.6
sip
six
smart-open
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
sqlparse==0.5.1
stack-data
starlette==0.41.3
statsmodels
streamlit
sympy==1.13.1
tables
tabulate
tblib
tenacity
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
toml
tomlkit
toolz
torch==2.0.1
torchvision==0.15.2
tornado
tqdm
traitlets
trimesh==3.13.1
truststore
Twisted
twisted-iocpsupport
typing_extensions
tzdata
tzlocal
uc-micro-py
ujson
Unidecode
uritemplate==4.1.1
urllib3
uvicorn==0.34.0
validators
w3lib
watchdog
watchfiles==1.0.3
wcwidth
webencodings==0.5.1
websocket-client
websockets==14.1
Werkzeug
whatthepatch
widgetsnbextension
win-inet-pton
wrapt
xarray
xlwings
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard==0.19.0